feat: return resized images with full search query
