+++ utils/index.ts
@@ -1,5 +1,6 @@
 import * as calendar from './calendar'
 import { logger } from './logs'
 import * as sql from './sql'
+import * as format from './format'
 
-export { calendar, logger, sql }
+export { calendar, logger, sql, format }
